alpaca_service/requirements.txt
alpaca-trade-api
psycopg2-binary
python-dotenv
# requests